hubble/relay: add possibility to provide a hostname to ClientConn
Update the ClientConnBuilder interface such that the ClientConn function
takes hostname as an additional parameter. This new parameter can be
ignored if not needed but is useful when establishing a secure
connection to verify a certificate's hostname.

As the ClientConnBuilder interface is updated, adapt existing
implementations for the change. The GRPCClientConnBuilder implementation
may not be provided with TLS configuration, in which case connection is
secured with TLS and the certificate is verified against the provided
hostname.

Signed-off-by: Robin Hahling <robin.hahling@gw-computing.net>